Create q5.java
A program which computes output according to alphabetical order value of letter. For example, hello (8,5,12,12,15) world (23,15,18,12,4) will be calculated as modulus (h-o)+(e-l)+(l), i.e., (15-8)+(12-5)+12 = 26 and world will be (w-d)+(o-l)+(r), i.e. (23-4)+(15-12)+18 = 40. Now for the int values 26 and 40 for hello world are concatenated to give output 2640.
Similarly "even deal"  will give 2612 via (14-5)+(22-5) = 26 and (12-4)+(5-1) = 12 after which 26 and 12 are concatenated and printed.